+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3628,7 +3628,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KN6QXC3P3FYJF6012JVBZHR0"),
+  id: z.string().default("01KN6WVRR137J1AWGWZ1N0R01R"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),