+++ packages/orq-rc/RELEASES.md
@@ -7989,3 +7989,13 @@ Based on:
 - [typescript v4.7.0-rc.42] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.42] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.42 - packages/orq-rc
+
+## 2026-04-02 10:50:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.0 (2.879.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.43] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.43] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.43 - packages/orq-rc